increased LoadSDF performance, added gzip support, and new tests
The SDF loader is now about 2.4x the original performance.

If the filename ends with ".gz" (case-insenitive) then the file
will be opened as a gzip file.

If a file object is passed in, then don't close it on exit - the
caller is responsible for that.

If the molecule is empty, return an empty data frame instead of None.