Rename file UnitTestDocTests in rdkitChem (#1263)
* 1194: Review assignments of range in Python code 

Task-Url: https://github.com/rdkit/rdkit/issues/1194
Either wrapped the range expression into a list or made sure that the
code is working with a range object.

* Rename file UnitTestDocTests in rdkitChem

When using a test runner, the files must have different names. The new
file UnitTestDocTests in rdkit.Chem is therefore renamed to
UnitTestDocTestsChem. 
Realized that this test was not included in test_list.py.