Adding a test case to make sure the sampling of isomers is deterministic.
Theoretically, the doc test should be doing this, but it didn't fail
in Python 3 like expected. I'm expecting this test case to fail on the
Travis Python 3 builds so I can grab the proper values from there to
insert.

Expecting the results to be the same for a given major version of
Python.