Improvements to GenerateStereoisomers (#1582)
* Improvements to GenerateStereoisomers:

- enumerate bond stereo as well
- change maxNumCenters to maxIsomers for more fine grained control
- allow sampling random subsets from potentially very large isomer spaces

Hoping this gets the 'easy' stuff out of the way before I submit the
more tricky stuff later at the UGM hack day: de-duplicating isomers;
and pseudo-stereo centers.

* Addressing comments in pull request 1582:

- Use six.moves.range instead of rolling own
- Adding double bond stereo to doc examples
- test cases for random seeding options
- test case for python OverflowError

* Make RDKit throw a ValueError when there is too many stereo centers and document the restriction.

* The random number generator is different in Python 3, so protect ourselves from that.

* Renamed GenerateStereoisomers to EnumerateStereoisomers and moved out of AllChem.py

* Getting doc tests to actually work.

* Port GenerateStereoisomers tests to EnumerateStereoisomers.

* Puts in a safety measure to make sure tryEmbedding can't cause
EnumerateStereoisomers to infinite loop.

* Test cases for the onlyUnassigned option.

* Removing imports no longer needed that EnumerateStereoisomers is located in EnumerateStereoisomers.py.

* Actually expecting this test case to fail between Python 2 and Python 3.

Maybe the '...' was preventing doc test from working?

* Adding a test case to make sure the sampling of isomers is deterministic.

Theoretically, the doc test should be doing this, but it didn't fail
in Python 3 like expected. I'm expecting this test case to fail on the
Travis Python 3 builds so I can grab the proper values from there to
insert.

Expecting the results to be the same for a given major version of
Python.