Replace link to incorrect thread
Under the *Neutralizing Charged Molecules* heading, there was a reference to the mailing list archive, however, this link pointed to an unrelated topic which related to the *Parallel conformation generation* heading. I have replaced it with the only thread from the mailing list archive which I could find which relates to the appropriate topic. Unfortunately, there does not appear to be the original code available from this thread as it point to a dead Google Code Wiki link. 

I apologize if I have included the incorrect thread, but I just wanted to try and fix this issue.