PgSQL: add support for PostgreSQL_CONFIG cmake var
This is useful for installations where pg_config is renamed to not
collide with the system default pg_config name.

We should drop 'find_package(PostgreSQL REQUIRED)' statement in
future because that's supposed to find PostgreSQL client library
to build against, not PostgreSQL server.  It might happen that
PostgreSQL server on the system is of different version than libpq
(Debian), but in future it will be problematic also on Fedora
(rhbz#1618698).

See also: https://gitlab.kitware.com/cmake/cmake/issues/17223