Avoid setting pandas max_colwidth to -1 (infinite)
When PandasTools is used, text columns in data frames can get extremely 
large because pandas global option display.max_width is set to -1. This 
also affects data frames without molecule images.

To circumvent this problem, utilize pandas TextAdjustment mechanism. 
These classes are intended to measure the length of text values in data 
frames. Provide a custom class "RenderHTMLAdjustment" that assigns a 
length of 0 if the given text is valid HTML. Consequently, pandas does 
not truncate these values. On the one hand, molecule images are not 
truncated (as before). On the other hand, other columns are not affected 
anymore and long text is abbreviated with dots.