Fix long columns pandas (#2655)
* Avoid setting pandas max_colwidth to -1 (infinite)

When PandasTools is used, text columns in data frames can get extremely 
large because pandas global option display.max_width is set to -1. This 
also affects data frames without molecule images.

To circumvent this problem, utilize pandas TextAdjustment mechanism. 
These classes are intended to measure the length of text values in data 
frames. Provide a custom class "RenderHTMLAdjustment" that assigns a 
length of 0 if the given text is valid HTML. Consequently, pandas does 
not truncate these values. On the one hand, molecule images are not 
truncated (as before). On the other hand, other columns are not affected 
anymore and long text is abbreviated with dots.

* Render svg molecule images correctly

If PandasTools.molRepresentation is set to svg and the data frame is 
wide enough, then only a part of the svg images is rendered. Fix this 
problem by specifying style properties and a viewbox.

Remove the method _get_svg_image since it is not used anywhere in rdkit.

* Only render RDKit molecule images in pandas data frames

Attach a data-content attribute to the HTML tags of generated molecule
images. Only render these tags as images and ignore other columns with
HTML or XML.

* Fix unit tests for PandasTools

Use df.to_html() instead of str(df) to test whether the data frame 
contains specific text, because the latter method truncates the result.

* Only render RDKit molecule images in pandas data frames - part 2

Setting escape to False leads to pandas rendering HTML in the whole data 
frame. Let pandas only escape molecule images by patching 
HTMLFormatter._write_cell.

* Implement fallback if pandas methods change

If some private methods in pandas are moved / deleted, use the patch of 
to_html that was used so far.

* Call old patch if the new patch causes exceptions