Fix #2854 by using utf-8 encoding in rdkit.Chem.Draw
Issue #2854 outlines a problem with the use of the `latin1` encoding since e.g. the bullet operator character (`\u2219`) can't be encoded with `latin1`. This commit introduces the `utf-8` encoding instead in order to be able to handle a larger character spectrum when encoding canvas text with `rdkit.Chem.Draw`.