Update build instructions for Linux
Numpy and matplotlib are not part of the base installation of miniconda
(anymore?). Add them to the list of packages that should be installed
using conda.

Specify the boost version (boost==1.67.0), because installing the most
recent version raises errors during "make" (at the moment).

Instruct the user to create a conda activate.d and deactivate.d script
to configure the path variables (LD_LIBRARIES_PATH, PYTHONPATH,
RDBASE).

Simplify the cmake instruction, because specifying BOOST_ROOT and
PYTHON_NUMPY_INCLUDE_PATH is not necessary anymore (in this setup).