FindMCS takes a long time when CompleteRingsOnly is set to True
I spent some time looking into issue #3965, and it seems like the MCS algorithm is stuck in a long (possibly infinite?) loop in certain queries where `BondCompareParameters.CompleteRingsOnly` is set to `True`. Interestingly, the algorithm does find the correct answer to the problem addressed in that issue after ~7s, but never completes the search. When `BondCompareParameters.CompleteRingsOnly` is set to `False`, the optimal solution is found in <1s.

I think part of the problem is that the algorithm isn’t terminating the ‘growing’ process of seeds that cannot become an MCS due to incomplete rings in `MaximumCommonSubgraph::growSeeds`.  These changes are my attempt to address this issue by only growing seeds that could potentially close incomplete rings, and finds the optimal MCS in ~.9s. However, this causes some other tests to fail so I think I am either missing something or this is not a good approach.

Does anyone who is more familiar with this code have any suggestions for an alternative approach?