Prefer wedging bonds that were wedged in the input file
If a molecule comes from SDF, it may have had input bonds
marked as wedged or dashed. When we choose wedge/dash bonds,
it would be friendly to take that into account.

I think that this would make it easier for users to visually
see that their input and output molecules from RDKit are
the same. Chirality is... somewhat tricky to perceive, so
any help we can give our users is nice.

The downside of this scheme is that two instances of the same
molecule that come out of RDK may have different bonds wedged/
dashed. I don't know if we were trying to canonicalize bond
wedging, though.