Fix LocaleSwitcher on Windows when RDK_BUILD_THREADSAFE_SSS not set
When `_WIN32` is defined but `RDK_BUILD_THREADSAFE_SSS` is not, the
`old_locale` variable is not initialized in the LocaleSwitcherImpl
constructor . This means that when the variable is consulted in the
destructor, it returns an empty (default constructed) string.

When `std::setlocale()` is called with an empty string, this causes
the locale to be set to "the user-preferred locale". This value is
determined from environment variables/system settings is not necessarily
the previous locale that was used by the program.

Best I can tell from PR#1892, the non-RDK_BUILD_THREADSAFE_SSS path was
simply missed when `old_locale` was added. Adding an analogous
initialization as in the RDK_BUILD_THREADSAFE_SSS-defined path should
cause both paths to behave similarly.