Fix exception in RegistrationHash for molecules with bad bond directions
The specific exception is:

    ..../site-packages/rdkit/Chem/EnumerateStereoisomers.py:329: in EnumerateStereoisomers
        Chem.SetDoubleBondNeighborDirections(isomer)
    E   RuntimeError: Pre-condition Violation
    E   	bad dir
    E   	Violation occurred on line 396 in file Code/GraphMol/Chirality.cpp
    E   	Failed Expression: dir == Bond::ENDUPRIGHT || dir == Bond::ENDDOWNRIGHT
    E   	RDKIT: 2022.09.3
    E   	BOOST: 1_76
    ----------------------------

This happens if we are enumerating stereoisomers and there is a
center with a bond direction between an enumerated center and a
double bond. This is a bad way to draw a bond wedge/dash, but
in Registration, we often see weird depictions straight from
some drug database.

Encountered at some customer site in Schrodinger software.