Remove `and` from C++ headers (#6003)
`and` is a C++ keyword, but it isn't supported by default on
some compilers (msvc). I'd like to disable it on _all_ platforms
in Schrodinger builds so that people don't accidentally post
code that fails on a single platform.

However, that means external headers need to use `&&` and `||`
instead of `and` and `or`, even in platform-specific code.