Add trans layout for double bonds in rings
If there is a trans double bond in a ring, it's important to
add a trans 2d depiction! This is important for visualization,
but it is also important for chemical accuracy. SDF uses the 2D
coordinates to discern bond stereochemistry, and some 3D embedding
techniques may start from 2D.

Without this commit, trans double bonds in rings invert on
roundtrip through SDF.

Here's a very simple assertion that fails before this commit
and passes after:

    smi = "C1=C/CCCCCCCCCCCC/1"
    assert smi == Chem.MolToSmiles(Chem.MolFromMolBlock(Chem.MolToMolBlock(Chem.MolFromSmiles(smi))))