Speed up GetPropsAsDict(), especially on mac
I noticed that we have some workflows for which
GetPropsAsDict() is a bottleneck. This function _should be_ pretty
fast, but exception stack unwinding is pretty slow, especially
on Mac. I'd expect conversion to int or double to _often_ fail
for properties with "string" type tag.

I created a molecule with 100 string properties and called
GetPropsAsDict() 3000 times.

On my M3 macbook pro...

Before this change: 100s
After this change:  0.12s

So, a 1000x speedup! That won't translate to a 1000x speedup
in the workflows I was looking up, but it definitely will change
the rate-limiting step.