Refactor smiles parsing code to use C++ lexer and parser (#7015)
This attempts to improve the SMILES parsing procedure to provide more
information about bad inputs by pointing to general location of the
offending token. The improved error messages currently only apply to syntax
errors, inputs with extra close parentheses, and inputs with too large numbers,
but this will provide a way to extend that in the future

Some examples of the improved error messages

| old version                            | new version            |
|----------------------------------------|------------------------|
|syntax error while parsing: c%()ccccc%() | syntax error while parsing input:<br /> c%()ccccc%() <br /> ---^ |
|syntax error while parsing: c%(100000)ccccc%(100000) | syntax error while parsing input: <br /> c%(100000)ccccc%(100000) <br /> --------^ |
|syntax error while parsing: COc(c1)cccc1C#|syntax error while parsing input: <br /> COc(c1)cccc1C# <br /> -------------^ |
|extra close parentheses while parsing: C) | extra close parentheses while parsing input: <br /> C) <br /> -^ |
| extra close parentheses while parsing: C1C)foo | extra close parentheses while parsing input: <br /> C1C)foo <br /> ---^ |
| number too large while parsing: [555555555555555555C] | number too large while parsing input: <br /> [555555555555555555C] <br /> -^ |

While this required using C++ lexers and parsers generated by
flex and bison, it doesn't affect the tokenization and parsing logic
as a whole. It does, however, improve thread safety for SMILES parsing
by removing usage of global states e.g. reducing the scope of the
`debugParse` parameter through removing the use of the `yysmiles_debug`
global.

The C++ lexer requires the FlexLexer.h header, which is installed by
flex, so I added it to the repository to support building with
RDK_USE_FLEXBISON turned off.