Adds the ranks of the neighboring atoms when running CIP labeling
I think this could be an optional additional annotation. I would
find this useful for replacing some schrodinger code that uses
the ranks. (I'd ultimately like to update that code to use
parity, but that's a much bigger project).

I was also hoping to do some performance work in the CIPLabeler,
and having this as a debug/validation tool would be nice.