Implements #TODO
This attempts to improve the SMILES parsing procedure to provide more
information about bad inputs by pointing to general location of the offending token.
The improved error messages currently only apply to syntax errors, inputs with
extra close parentheses, and inputs with too large numbers, but this will
provide a way to extend that in the future

Some examples of the improved error messages

| old version                            | new version            |
|----------------------------------------|------------------------|
|syntax error while parsing: c%()ccccc%() | syntax error while parsing input:<br /> `c%()ccccc%()` <br /> `~~~^` |
|syntax error while parsing: c%(100000)ccccc%(100000) | syntax error while parsing input: <br /> `c%(100000)ccccc%(100000)` <br /> `~~~~~~~~^` |
|syntax error while parsing: COc(c1)cccc1C#|syntax error while parsing input: <br /> `COc(c1)cccc1C#` <br /> `~~~~~~~~~~~~~^` |
|extra close parentheses while parsing: C) | extra close parentheses while parsing input: <br /> `C)` <br /> `~^` |
| extra close parentheses while parsing: C1C)foo | extra close parentheses while parsing input: <br /> `C1C)foo` <br /> `~~~^` |
| number too large while parsing: [555555555555555555C] | number too large while parsing input: <br /> `[555555555555555555C]` <br /> `~~~~~~~~~~^` |

This was achieves by extending the lexing and parsing procedure to track
the current token position with a new variable, so the reported position
may not be 100% accurate at all times but should be helpful in reducing
the amount of work done to find the bad location.

Additionally, I updated the build instructions to strip the #line macros
from the generated C++ files, which required a modern flex/bison
versions.