Ignore invalid chirality labels when reading MAE inputs
For a host of reasons, inputs chirality labels can be invalid as we try
to read an input MAE file or buffer. This can be annoying for situations
when we have other means of calculating the stereochemistry information,
so we should try to ignore bad labels and let the user deal with the
"bad" stereochemistry.