Fix default wiggly bond writing behaviour in CXSMILES writer (#8348)
When writing CXSMILES outpupts with the RestoreBondDirOptionClear
argument, which is enabled by default, we fail to write wiggly bond
information. I traced the issue to us clearing said information on the relevant
bonds during the pre-processing stage, so I fixed fixed the issue by
removing the that logic for wiggly bonds.