Ignore invalid chirality labels when reading MAE inputs (#8347)
* Ignore invalid chirality labels when reading MAE inputs

For a host of reasons, inputs chirality labels can be invalid as we try
to read an input MAE file or buffer. This can be annoying for situations
when we have other means of calculating the stereochemistry information,
so we should try to ignore bad labels and let the user deal with the
"bad" stereochemistry.

* remove stereo bond label changes

* Remove debugging code

* restore whitespace

* remove redundant &

* don't clear all stereo when we see invalid chirality labels

* use catch2 generators for test

* replace some requires with check