Enable ccache to speed up local clean builds (#8613)
This enables usage of ccache to cache compiler outputs when it's installed on
the machine. This will allow us to skip having to re-compile unchanged files
during clean builds. A no-op clean build on my intel mac goes from taking 40mins
to 3mins with `make -j10` and an up-to-date cache.