Fixed a problem writing pickle data where string query data wasn't correctly handled
The pickle writer was outputting a "RDVal" instead of a string, and there wasn't
any method to serialize the RDVal. By correctly getting the string, it was able
to be serialized properly, fixing the pickle, which fixed the unpickling, which
threw an error.

We also added a check for pickleFlags, but this wasn't necessary for the
catch_pickle tests, but maybe elsewhere it will be relevant.